+++ src/modules/session/dto/refresh-session.dto.ts
@@ -1,0 +1,3 @@
+export class RefreshSessionDto {
+    refreshToken: string;
+}
